style: Make a structure of project: add include and src directories for headers and source files.
chore: Create main file.

chore: Create input_handler and input_checker files with functions to
handle with correct and incorrect inputs.

chore: Include structs of forks, philos and general data of the program.

chore: Create mutex_handler and thread_handler files with functions to
deal with mutexes and threads and handle it's errors.